chore : 바우처 생성 Validity 체크하는 로직 voucher Service로 변경

+++ kdtSpringOrder/src/main/java/org/prgrms/kdt/dto/VoucherSaveRequestDto.java
@@ -5,9 +5,9 @@ import org.prgrms.kdt.enums.VoucherType;
 public class VoucherSaveRequestDto {
 
     private VoucherType voucherType;
-    private int discount;
+    private long discount;
 
-    public VoucherSaveRequestDto(VoucherType voucherType, int discount) {
+    public VoucherSaveRequestDto(VoucherType voucherType, long discount) {
         this.voucherType = voucherType;
         this.discount = discount;
     }
@@ -16,7 +16,7 @@ public class VoucherSaveRequestDto {
         return voucherType;
     }
 
-    public int getDiscount() {
+    public long getDiscount() {
         return discount;
     }
